Add files via upload
This is the code for actually training the neural network. This will run and train the CNN, and will save all of the weights after it has finished. I also included my saved model in this repository. I'm adding a small tool so that you can run an image of your choice through the network, it's almost finished.